+++ modules/features/widgets/src/main/kotlin/au/com/shiftyjelly/pocketcasts/widget/ui/LargePlayer.kt
@@ -2,7 +2,6 @@ package au.com.shiftyjelly.pocketcasts.widget.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.glance.GlanceModifier
 import androidx.glance.GlanceTheme
 import androidx.glance.LocalContext
@@ -12,13 +11,10 @@ import androidx.glance.background
 import androidx.glance.layout.Alignment
 import androidx.glance.layout.Column
 import androidx.glance.layout.Spacer
-import androidx.glance.layout.fillMaxSize
 import androidx.glance.layout.fillMaxWidth
 import androidx.glance.layout.height
 import androidx.glance.layout.padding
-import androidx.glance.text.FontWeight
-import androidx.glance.text.Text
-import androidx.glance.text.TextStyle
+import androidx.glance.layout.wrapContentHeight
 import au.com.shiftyjelly.pocketcasts.widget.action.OpenPocketCastsAction
 import au.com.shiftyjelly.pocketcasts.widget.data.LargePlayerWidgetState
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
@@ -31,7 +27,7 @@ internal fun LargePlayer(state: LargePlayerWidgetState) {
         Column(
             modifier = GlanceModifier
                 .fillMaxWidth()
-                .height(350.dp)
+                .wrapContentHeight()
                 .cornerRadius(6.dp)
                 .background(GlanceTheme.colors.primaryContainer)
                 .padding(16.dp),
@@ -47,29 +43,28 @@ internal fun LargePlayer(state: LargePlayerWidgetState) {
                     useDynamicColors = state.useDynamicColors,
                 )
             } else {
-                val secondaryTextColor = if (state.useDynamicColors) {
-                    GlanceTheme.colors.onPrimaryContainer
-                } else {
-                    GlanceTheme.colors.onSecondaryContainer
-                }
-
                 Column(
                     verticalAlignment = Alignment.CenterVertically,
                     horizontalAlignment = Alignment.CenterHorizontally,
-                    modifier = GlanceModifier.fillMaxSize().clickable(OpenPocketCastsAction.action()),
+                    modifier = GlanceModifier
+                        .fillMaxWidth()
+                        .height(190.dp)
+                        .clickable(OpenPocketCastsAction.action()),
                 ) {
-                    Text(
+                    NonScalingText(
                         text = LocalContext.current.getString(LR.string.widget_nothing_in_up_next),
-                        maxLines = 1,
-                        style = TextStyle(color = GlanceTheme.colors.onPrimaryContainer, fontSize = 15.sp, fontWeight = FontWeight.Bold),
+                        textSize = 16.dp,
+                        useDynamicColors = state.useDynamicColors,
+                        isBold = true,
                     )
                     Spacer(
                         modifier = GlanceModifier.height(4.dp),
                     )
-                    Text(
+                    NonScalingText(
                         text = LocalContext.current.getString(LR.string.widget_check_out_discover),
-                        maxLines = 1,
-                        style = TextStyle(color = secondaryTextColor, fontSize = 13.sp),
+                        textSize = 13.dp,
+                        useDynamicColors = state.useDynamicColors,
+                        alpha = 0.8,
                     )
                 }
             }